Сделал модуляцию, дз 5.
